Add travis checkage for unstaged files

+++ scripts/gh-build.sh
@@ -8,7 +8,10 @@ git config --global user.name "Travis CI"
 # build
 npm run build
 
-# deploy
-git add .
-git commit -m "Build revision" &> /dev/null
-git push --force "https://${GITHUB_TOKEN}@github.com/${GITHUB_REPO}.git" HEAD:master > /dev/null 2>&1
+if [ -n "$(git status --porcelain)" ]; then
+	git add .
+	git commit -m "Build revision"
+	git push --force "https://${GITHUB_TOKEN}@github.com/${GITHUB_REPO}.git" HEAD:master > /dev/null 2>&1
+else
+ exit 0
+fi
